requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
cmake==3.31.6
colorama==0.4.6
deprecation==2.1.0
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
multidict==6.1.0
numpy==2.2.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pillow==11.1.0
postgrest==0.19.3
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.1
rich==13.9.4
rich-toolkit==0.13.2
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
yarl==1.18.3
